portal/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
Django==3.2.6
django-ajax-datatable==4.2.1
django-ajax-tables==1.1.1
django-cart==1.0.4
django-cors-headers==3.8.0
django-filter==2.4.0
django-online-users==0.3
django-tables2==2.4.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
iommi==3.1.1
msgpack==1.0.2
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
tri.declarative==5.7.0
tri.struct==4.1.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
websocket-client==1.2.1
zope.interface==5.4.0
